requirements.txt
boto3==1.34.7
apscheduler==3.10.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
sqlalchemy==2.0.23